+++ custom_components/solarcharger/modules/coordinator.py
@@ -103,6 +103,7 @@ class SolarChargerCoordinator(ScOptionState):
         self._net_power_update_count: int = 0
 
         # Started tracking weather
+        self._weather_provider: str = None
         self._tracking_weather: bool = False
 
     # ----------------------------------------------------------------------------
@@ -531,25 +532,42 @@ class SolarChargerCoordinator(ScOptionState):
             self._hass.async_create_task(self._async_update_weather_sensor(entity_id))
 
     # ----------------------------------------------------------------------------
+    def _subscribe_weather(self, weather_provider: str) -> None:
+        """Subscribe weather updates."""
+
+        if self._tracker.track_weather_update(self._async_handle_weather_update):
+            # Populate weather sensor with data.
+            self._async_handle_weather_update(None)
+            self._weather_provider = weather_provider
+            self._tracking_weather = True
+        else:
+            self._weather_provider = None
+            self._tracking_weather = False
+
+    # ----------------------------------------------------------------------------
+    def _unsubscribe_weather(self) -> None:
+        """Unsubscribe weather updates."""
+
+        self._tracker.untrack_weather_update()
+        self._update_sensor_attribute(SENSOR_WEATHER_FORECAST, STATE_UNKNOWN, None)
+        self._weather_provider = None
+        self._tracking_weather = False
+
+    # ----------------------------------------------------------------------------
     def _check_weather_provider(self) -> None:
         """Track weather if weather provider is defined."""
 
         entity_id = self.get_weather_provider()
         if entity_id is not None:
+            if entity_id != self._weather_provider and self._tracking_weather:
+                self._unsubscribe_weather()
+
             if not self._tracking_weather:
-                if self._tracker.track_weather_update(
-                    self._async_handle_weather_update
-                ):
-                    # Populate weather sensor with data.
-                    self._async_handle_weather_update(None)
-                    self._tracking_weather = True
+                self._subscribe_weather(entity_id)
+
         else:
             if self._tracking_weather:
-                self._tracker.untrack_weather_update()
-                self._update_sensor_attribute(
-                    SENSOR_WEATHER_FORECAST, STATE_UNKNOWN, None
-                )
-                self._tracking_weather = False
+                self._unsubscribe_weather()
 
     # ----------------------------------------------------------------------------
     # Periodic functions